Leonardo

Documentation for Leonardo.

Leonardo.AbstractArrow

Leonardo.AbstractDirection

Leonardo.AbstractEnd

Leonardo.AbstractFreeChar

Leonardo.AbstractLineSize

Leonardo.AbstractLineType

Leonardo.Bar

Leonardo.BlackArrow

Leonardo.BlackRectangle

Leonardo.Canvas

Leonardo.ConnectorStyle

Leonardo.Dashed1

Leonardo.Dashed2

Leonardo.Dashed3

Leonardo.Double

Leonardo.Down

Leonardo.EndStyle

Leonardo.Half1

Leonardo.Half2

Leonardo.Heavy

Leonardo.Horizontal

Leonardo.LargeDot

Leonardo.Left

Leonardo.Light

Leonardo.LineStyle

Leonardo.Lozenge

Leonardo.NoDirection

Leonardo.NoEnd

Leonardo.NoLine

Leonardo.PathStyle

Leonardo.PathStyle

Leonardo.PathStyle

Leonardo.PathStyle

Leonardo.PrintStyle

Leonardo.Right

Leonardo.SmallDot

Leonardo.Solid

Leonardo.Up

Leonardo.Vertical

Leonardo.WhiteArrow

Leonardo.WhiteRectangle

Leonardo.biconnector

Leonardo.drawconnector!

Leonardo.drawend!

Leonardo.drawtext!

Leonardo.drawtextbox!

Leonardo.quadriconnector

Leonardo.triconnector

Leonardo.AbstractArrow — Type

Abstract type defining an arrow head character.

Leonardo.AbstractDirection — Type

Abstract type defining a direction, used to define the surroundings of a ConnectorStyle, to define the EndStyle to draw paths, and to define the BoxStyle to draw boxes.

Leonardo.AbstractEnd — Type

Abstract type defining an ending character for EndStyle and PathStyle.

Leonardo.AbstractFreeChar — Type

Abstract type defining a free character, i.e. a character without any surrounding informations (no line size, no direction).

Leonardo.AbstractLineSize — Type

Abstract type defining the drawing size of a line. Subtypes are:

NoLine

Light

Heavy

Double

Leonardo.AbstractLineType — Type

Abstract type defining the type of line. Subtypes are:

Solid

Dashed1

Dashed2

Dashed3

Half1

Half2

Leonardo.Bar — Type

Type defining a bar character for an EndStyle.

Example

julia> Leonardo.char(Bar, Left, Heavy)
'┣': Unicode U+2523 (category So: Symbol, other)

Leonardo.BlackArrow — Type

Type defining a black arrow head character for an EndStyle.

Example

julia> Leonardo.char(BlackArrow, Left)
'◂': Unicode U+25C2 (category So: Symbol, other)

Leonardo.BlackRectangle — Type

Type defining a black rectangle character.

Example

julia> Leonardo.char(BlackRectangle)
'▮': Unicode U+25AE (category So: Symbol, other)

Leonardo.Canvas — Type

Canvas(width, height; [bgchar = ' '], [defstyle = PrintStyle()])

Define a drawing area of size widthxheight, where width (resp. height) is the number of characters along the X (resp. Y) axis. bgchar is the background character by default it is a space. defstyle is the default PrintStyle used to print this Canvas. The origin of coordinate system is the upper left character:

┌─────────────────────────────────────────────────────────────────────────────▸X
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
│∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙∙
▾
Y

Leonardo.ConnectorStyle — Type

ConnectorStyle(left, up, right, down)

Type defining a connector character. left, up, right and down are line styles surrounding the connector.

See biconnector, triconnector and quadriconnector for easy creation.

Leonardo.Dashed1 — Type

Type defining a dashed line.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light, Dashed1))
┄┄┄┄┄┄┄┄┄┄

Leonardo.Dashed2 — Type

Type defining a dashed line.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light, Dashed2))
┈┈┈┈┈┈┈┈┈┈

Leonardo.Dashed3 — Type

Type defining a dashed line.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light, Dashed3))
╌╌╌╌╌╌╌╌╌╌

Leonardo.Double — Type

Type defining a double sized line.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Double))
══════════

Leonardo.Down — Type

Type defining a down direction, used to define the surroundings of a ConnectorStyle, to define the EndStyle to draw paths, and to define the BoxStyle to draw boxes.

Leonardo.EndStyle — Type

EndStyle(etype, edir, lstyle)

Create a type defining a line ending character, mostly used in internal to draw path. etype is one of NoEnd, BlackArrow, WhiteArrow or any subtypes o f AbstractFreeChar. edir gives the direction of ending character. lstyle is the style of the line for which ending character must be drawn.

See drawend! for usage.

Leonardo.Half1 — Type

Type defining a half line type at the left for horizontal lines, or at the top for vertical lines.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light, Half1))
╴╴╴╴╴╴╴╴╴╴

Leonardo.Half2 — Type

Type defining a half line type at right left for horizontal lines, or at the bottom for vertical lines.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light, Half2))
╶╶╶╶╶╶╶╶╶╶

Leonardo.Heavy — Type

Type defining a heavy sized line.

Example

ulia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Heavy))
━━━━━━━━━━

Leonardo.Horizontal — Type

Type defining a line with a horizontal orientation. See drawline! or drawpath! for usage.

Leonardo.LargeDot — Type

Type defining a large dot character.

Example

julia> Leonardo.char(LargeDot)
'•': Unicode U+2022 (category Po: Punctuation, other)

Leonardo.Left — Type

Type defining a left direction, used to define the surroundings of a ConnectorStyle, to define the EndStyle to draw paths, and to define the BoxStyle to draw boxes.

Leonardo.Light — Type

Type defining a light sized line.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light))
──────────

Leonardo.LineStyle — Method

LineStyle(lsize, ltype)

Create a LineStyle type

Leonardo.Lozenge — Type

Type defining a lozenge character.

Example

julia> Leonardo.char(Lozenge)
'◊': Unicode U+25CA (category So: Symbol, other)

Leonardo.NoDirection — Type

Type defining no specific direction, only used by EndStyle for mutiple dispatch convenience.

Leonardo.NoEnd — Type

Type used to define an EndStyle that will draw a normal line character.

Leonardo.NoLine — Type

Type defining a missing line, mainly used by 'ConnectorStyle'.

Leonardo.PathStyle — Method

PathStyle(lstyle1, lstyle2, etype1, etype2)

Create a PathStyle type used by drawpath! to draw paths.

Leonardo.PathStyle — Method

PathStyle(lstyle, etype1, etype2)

Create a PathStyle type used by drawpath! to draw paths.

Leonardo.PathStyle — Method

PathStyle(etype1, etype2)

Create a PathStyle type used by drawpath! to draw paths.

Leonardo.PathStyle — Method

PathStyle(lstyle, etype)

Create a PathStyle type used by drawpath! to draw paths.

Leonardo.PrintStyle — Type

PrintStyle(; kwargs...)

Store keyword arguments paased to the printstyled function:

bold: false by default

color: :normal by default

blink: false by default

reverse: false by default

hidden: false by default

underline: false by default

For further details see printstyled documentation.

Leonardo.Right — Type

Type defining a right direction, used to define the surroundings of a ConnectorStyle, to define the EndStyle to draw paths, and to define the BoxStyle to draw boxes.

Leonardo.SmallDot — Type

Type defining a small dot character.

Example

julia> Leonardo.char(SmallDot)
'∙': Unicode U+2219 (category Sm: Symbol, math)

Leonardo.Solid — Type

Type defining a solid line.

Example

julia> c = Canvas(10, 1);

julia> drawline!(c, (1,1), 10, Horizontal, LineStyle(Light, Solid))
──────────

Leonardo.Up — Type

Type defining an up direction, used to define the surroundings of a ConnectorStyle, to define the EndStyle to draw paths, and to define the BoxStyle to draw boxes.

Leonardo.Vertical — Type

Type defining a line with a vertical orientation. See drawline! or drawpath! for usage.

Leonardo.WhiteArrow — Type

Type defining a white arrow head character for an EndStyle.

Example

julia> Leonardo.char(WhiteArrow, Left)
'◃': Unicode U+25C3 (category So: Symbol, other)

Leonardo.WhiteRectangle — Type

Type defining a white rectangle character.

Example

julia> Leonardo.char(WhiteRectangle)
'▯': Unicode U+25AF (category So: Symbol, other)

Leonardo.biconnector — Function

biconnector(ldir1, lstyle1, ldir2, lstyle2)

Create a ConnectorStyle between two lines with line styles lstyle1 and lstyle2 and with positions ldir1 and ldir2.

Example

julia> c = Canvas(10, 1);

julia> l1 = LineStyle(Light);

julia> l2 = LineStyle(Heavy, Half1);

julia> drawline!(c, (1,1), 4, Horizontal, l1)
────

julia> drawline!(c, (6,1), 4, Horizontal, l2)
──── ╸╸╸╸

julia> bc = biconnector(Left, l1, Right, l2);

julia> drawconnector!(c, (5,1), bc)
────╼╸╸╸╸

Leonardo.drawconnector! — Method

drawconnector!(canvas, (x,y), cstyle, [prstyle = defstyle(canvas)])

Draw a connector at location (x,y), with a connector style cstyle.

See biconnector, triconnector and quadriconnector for usage examples.

Leonardo.drawend! — Method

drawend!(canvas, (x, y), estyle, prstyle = defstyle(canvas)

Draw an end character for a line defined by estyle a EndStyle type.

Example

julia> c = Canvas(10, 1);

julia> es = EndStyle(NoEnd, Left, LineStyle());

julia> drawend!(c, (1, 1), es)
─

julia> es = EndStyle(BlackArrow, Left, LineStyle());

julia> drawend!(c, (1, 1), es)
◂

julia> es = EndStyle(Bar, Left, LineStyle());

julia> drawend!(c, (1, 1), es)
├

julia> es = EndStyle(Lozenge, NoDirection, LineStyle(NoLine));

julia> drawend!(c, (1, 1), es)
◊

Leonardo.drawtext! — Method

drawtext!(canvas, (x,y), text, ori = Horizontal, prstyle = defstyle(canvas))

Draw a string text string at location (x,y). ori indicates the text orientation, either Horizontal or Vertical. prstyle defines the print style for REPL rendering.

Example

julia> c = Canvas(80, 15);

julia> title = """
       So Long,
       and Thanks for All the Fish""";

julia> drawtext!(c, (5,10), title, Horizontal);

julia> ps = PrintStyle(color = :red, blink = true);

julia> drawtext!(c, (60,1), "Mostly Harmless", Vertical, ps);

julia> c
                                                           M
                                                           o
                                                           s
                                                           t
                                                           l
                                                           y

                                                           H
                                                           a
    So Long,                                               r
    and Thanks for All the Fish                            m
                                                           l
                                                           e
                                                           s
                                                           s

Leonardo.drawtextbox! — Method

drawtextbox!(canvas, (x,y), text, ori = Horizontal, bstyle = BoxStyle(), prstyle = defstyle(canvas))

Similar to [drawtext!], but with a bounding box defined by the boxstyle bstyle.

See also drawtext! and BoxStyle.

Example

julia> c = Canvas(80, 8);

julia> bs = BoxStyle(LineStyle(Heavy));

julia> h2g2 = """The Hitchhiker's Guide
       to the Galaxy""";

julia> drawtextbox!(c, (5,5), h2g2, Horizontal, bs)

    ┎──────────────────────┒
    ┃The Hitchhiker's Guide┃
    ┃to the Galaxy         ┃
    ┖──────────────────────┚

Leonardo.quadriconnector — Function

quadriconnector(ldir1, lstyle1, ldir2, lstyle2, ldir3, lstyle3, ldir4, lstyle4)

Create a ConnectorStyle between four lines with line styles lstyle1, lstyle2, lstyle3, and lstyle4 and with positions ldir1, ldir2, ldir3 and ldir4.

Example

julia> c = Canvas(10, 5);

julia> l1 = LineStyle(Double);

julia> l2 = LineStyle(Double);

julia> l3 = LineStyle(Dashed1);

julia> l4 = LineStyle(Dashed3);

julia> drawline!(c, (1,3), 4, Horizontal, l1);

julia> drawline!(c, (6,3), 4, Horizontal, l2);

julia> drawline!(c, (5,4), 2, Vertical, l3);

julia> drawline!(c, (5,2), -2, Vertical, l4);

julia> c
    ╎
    ╎
════ ════
    ┆
    ┆

julia> qc = quadriconnector(Left, l1, Right, l2, Down, l3, Up, l4);

julia> drawconnector!(c, (5,3), qc)
    ╎
    ╎
════╪════
    ┆
    ┆

Leonardo.triconnector — Function

triconnector(ldir1, lstyle1, ldir2, lstyle2, ldir3, lstyle3)

Create a ConnectorStyle between three lines with line styles lstyle1, lstyle2 and lstyle3 and with positions ldir1, ldir2 and ldir3.

Example

julia> c = Canvas(10, 3);

julia> l1 = LineStyle(Double);

julia> l2 = LineStyle(Double);

julia> l3 = LineStyle();

julia> drawline!(c, (1,1), 4, Horizontal, l1);

julia> drawline!(c, (6,1), 4, Horizontal, l2);

julia> drawline!(c, (5,2), 2, Vertical, l3);

julia> c
════ ════
    │
    │

julia> tc = triconnector(Left, l1, Right, l2, Down, l3);

julia> drawconnector!(c, (5,1), tc)
════╤════
    │
    │